requirements.txt
autopep8==2.1.0
beautifulsoup4==4.12.4
black==24.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==22.0.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
reportlab==4.3.1
python-dotenv==1.0.1
